+++ example/readme_example.dart
@@ -7,12 +7,17 @@ import 'readme_example.reflectable.dart';
 @SmartArg.reflectable
 @Parser(description: 'Hello World application')
 class Args extends SmartArg {
-  @StringArgument(help: 'Name of person to say hello to')
+  @StringArgument(
+    help: 'Name of person to say hello to',
+    //Environment Variable will be used if defined and not otherwise specified
+    environmentVariable: "GREETING_NAME",
+  )
   String name = 'World'; // Default to World
 
   @StringArgument(
     help: 'Message to say to person',
     mustBeOneOf: ['Hello', 'Goodbye'],
+    environmentVariable: 'GREETING_TYPE',
   )
   String greeting = 'Hello'; // Default to Hello
 
@@ -21,6 +26,7 @@ class Args extends SmartArg {
     isRequired: true,
     minimum: 1,
     maximum: 100,
+    environmentVariable: 'GREETING_COUNT',
   )
   late int count;